requirements.txt
pandas>=2.0
numpy>=1.24
plotly>=5.20
streamlit>=1.34
scikit-learn>=1.4
sentence-transformers>=2.7
transformers>=4.40
torch>=2.2
python-dotenv>=1.0
google-genai>=1.0
